refactor, do not store unmanaged fields
